+++ som-interpreter-ast/src/primitives/object.rs
@@ -40,7 +40,7 @@ fn class(universe: &mut Universe, args: Vec<Value>) -> Return {
 fn object_size(_: &mut Universe, _: Vec<Value>) -> Return {
     const _: &'static str = "Object>>#objectSize";
 
-    Return::Local(Value::Integer(std::mem::size_of::<Value>() as i64))
+    Return::Local(Value::Integer(std::mem::size_of::<Value>() as i32))
 }
 
 fn hashcode(_: &mut Universe, args: Vec<Value>) -> Return {
@@ -52,7 +52,7 @@ fn hashcode(_: &mut Universe, args: Vec<Value>) -> Return {
 
     let mut hasher = DefaultHasher::new();
     value.hash(&mut hasher);
-    let hash = (hasher.finish() as i64).abs();
+    let hash = (hasher.finish() as i32).abs();
 
     Return::Local(Value::Integer(hash))
 }
